Make the preprocessors vector a field in the ReplicaFactory class
ReplicaImp: reset smart pointer instead of explicit delete/release

Remove duplicated preProcessor stop mechanism

Threadlocaldata removal
