+++ kvbc/include/Replica.h
@@ -34,6 +34,7 @@
 #include "log/logger.hpp"
 #include "kvbc_app_filter/value_from_kvbc_proto.h"
 #include "newest_public_event_group_record_time.h"
+#include "ReplicaFactory.hpp"
 
 namespace concord::kvbc {
 
@@ -244,7 +245,7 @@ class Replica : public IReplica,
   std::shared_ptr<storage::IDBClient> m_metadataDBClient;
   std::unique_ptr<bft::communication::ICommunication> m_ptrComm;
   const bftEngine::ReplicaConfig &replicaConfig_;
-  bftEngine::IReplica::IReplicaPtr m_replicaPtr = nullptr;
+  bftEngine::ReplicaFactory::IReplicaPtr m_replicaPtr;
   std::shared_ptr<ICommandsHandler> m_cmdHandler = nullptr;
   bftEngine::IStateTransfer *m_stateTransfer = nullptr;
   concord::storage::DBMetadataStorage *m_metadataStorage = nullptr;
